added feature allowing admin to delete another user's profile + tests + bug fixes

+++ src/Service/User/UserService.php
@@ -3,17 +3,26 @@
 namespace App\Service\User;
 
 use App\Entity\User;
+use App\Enum\UserRoleEnum;
+use App\Exception\AdminDeletionException;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
 
 class UserService
 {
-    public function __construct(private EntityManagerInterface $entityManager, private TokenStorageInterface $tokenStorage)
+    public function __construct(
+        private readonly EntityManagerInterface $entityManager,
+        private readonly TokenStorageInterface  $tokenStorage
+    )
     {
     }
 
     public function deleteUser(User $user): void
     {
+        if (in_array(UserRoleEnum::ROLE_ADMIN, $user->getRoles())) {
+            throw new AdminDeletionException('You can not delete users with admin role', 403);
+        }
+
         $this->entityManager->remove($user->getUserProfile());
         $this->entityManager->remove($user);
